Update Sort Files in Folders.py
This was the first version where I wanted to increase the scope to add a current folder. I started thinking of how the Downloads folder is like the port of the networked PC. It is very different for the Phone though. Phones might be more like semi-permeable membrane than a land boundary.