̸f̴r̵o̵m̵ ̶d̷a̶t̶e̶t̸i̸m̶e̴ ̴i̷m̷p̷o̴r̶t̴ ̷d̷a̷t̴e̵t̵i̴m̷e̸,̶ ̴t̴i̷m̶e̴d̸e̷l̷t̶a̵ Its still broken
But i am in a new rabbithole
Non-Linear Video Editor
I want to connect a few nodes around and have a video edited. 
i am amazed at the shitpost potential. and I scurry around for the materials.
- one Olive2.0 but it should be an experimental alpha build that constantly crashes
- one prasar Bharati video from the Prasar Bharati Video Archives on YT

I am now obsessed with the podcast - all of them are in queue in 1.7x