i am done for the night
i have thought of how i am going to upload the code to github when I am done. I want to tell it like a story though. I start thinking of how interesting it would be to have chatgpt and its responses as a branched commits kind of structure since I am going and editing my questions earlier in the thread. but then that thread is lost. all memory of it is gone. I am always so surprised when the feeling of loss kicks in because you tried looking for something that you erased to ask a new question.