2:30 pm the next day
i woke up, and spoke to S for three hours. I spoke to them of  all the weird twists and turns of this story. And they say, you should publish it. 

Something is happening in my brain as I talk to them

we speak of public spaces, internet as a public space.
How loitering in a public space means claiming it as your own. They tell me about a group of women who gather to go loiter at night and do all the things that people tell them not to do. Like sleep outside at night. We talk about embodied Data and how we wield it as a tool, id card, something we care about, something we pay for, and something we pay with. If the internet is a human right/ What does it mean for the Internet to be a human right. I remember this article 
https://www.indiatoday.in/india/story/kashmir-survived-without-4g-internet-for-months-with-coronavirus-it-really-needs-it-1659883-2020-03-26

I tell them about my theory on what is going to happen if AI is the future. 

"I think the general public will have very little reason to remain on the internet after sometime. The creator economy will be devastated. what happens the post-attention economy doesn't look too good. The only entities with any need to be on the internet will be these AI agents, A new paradigm but reminiscent of the platform paradigm of yesteryear. i see class systems for agents like bollywood tabloids calling random actors A-tier and B-tier. It is no long a public space, but a giant TV, ENDLESS HEDONY"