as i get closer to the end, i get madder. Rage against the Machine, literally
apparently chatgpt doesnt understand that 

    if datetime.now() - file_modification_time > threshold:

is the exact opposite of what I want. Its been 18 hours and 20+ versions of this till now. Not once has it told me that this might be an issue. Or did it just change somewhere in between those hours and versions. Was this a hallucination from the Machine or from me?