5pm
OH IT IS GLORIOUS! IT WORKS!

But now I don't understand this bit of code. I ask the fkin machine to give it some comments line by line. but by now it is saturated in 20 hours of me asking the same things so it keeps returning comments as broad-level guides and not as specific things. I like to think the machine is burnt out.

I test this version and all the excitement I had to put it to work on the downloads folder is gone. it is now 5 pm. i am tired. but my brain is buzzing.