am i making repo art rn?
its been 4 hours since the code was taken out of the oven. I have uploaded every version to Github. 

I planned to tell the story on the commits page and through the interface elements of Github. I also used ASCII art and typographic elements to bake the story into the interface. the code changes behind the commit might be the fun bit for you. 

I am going to use hypothesis(https://hyp.is) as a milestone marker and traffic signal for this narrative. 

Probably set up the Wiki as a shady backhouse of this seemingly innocuous script repository that people looking for a folder sorter will not find. 

micahchoo out.